+++ client/CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `QueryParameters::empty()`;
 - MQTT `Error`;
+- `Message::HashTransactionEssence`;
+- `Response::TransactionEssenceHash`;
 
 ### Changed